Move IconColumn to itemviews/columns
- it is only used there (and in tests)
