+++ picard/ui/columns.py
@@ -116,42 +116,6 @@ class ImageColumn(Column):
         raise NotImplementedError
 
 
-class IconColumn(ImageColumn):
-    _header_icon = None
-    header_icon_func = None
-    header_icon_size = QtCore.QSize(0, 0)
-    header_icon_border = 0
-
-    @property
-    def header_icon(self):
-        # icon cannot be set before QApplication is created
-        # so create it during runtime and cache it
-        # Avoid error: QPixmap: Must construct a QGuiApplication before a QPixmap
-        if self._header_icon is None:
-            self._header_icon = self.header_icon_func()
-        return self._header_icon
-
-    def set_header_icon_size(self, width, height, border):
-        self.header_icon_size = QtCore.QSize(width, height)
-        self.header_icon_border = border
-        self.size = QtCore.QSize(width + 2*border, height + 2*border)
-        self.width = self.size.width()
-
-    def paint(self, painter, rect):
-        icon = self.header_icon
-        if not icon:
-            return
-        h = self.header_icon_size.height()
-        w = self.header_icon_size.width()
-        border = self.header_icon_border
-        padding_v = (rect.height() - h) // 2
-        target_rect = QtCore.QRect(
-            rect.x() + border, rect.y() + padding_v,
-            w, h
-        )
-        painter.drawPixmap(target_rect, icon.pixmap(self.header_icon_size))
-
-
 class Columns(MutableSequence):
     def __init__(self, iterable=None, default_width=None):
         self._list = list()
